refactor: make tx-completion capacity an Indexer::new arg, move lookup_tx_completion to tx.rs
Indexer::new now takes tx_completion_capacity directly (1024 in
production) instead of a separate with_channel_capacity constructor.
