+++ src/tx.rs
@@ -1,15 +1,19 @@
 use std::collections::{BTreeMap, HashMap};
+use std::sync::Arc;
 
-use anyhow::Result;
+use anyhow::{bail, Error, Result};
 use edn::kw;
 use edn::symbols::Keyword;
+use slatedb::DbReadOps;
 
 use crate::codec::{self, encode_datatype, encode_i64_bytes, Encode};
-use crate::indexer::vae_key_to_parts;
+use crate::indexer::{ave_key_to_parts, eav_key_to_parts, vae_key_to_parts, TxCompletion};
 use crate::iterator::slate_key_iterator::SlateKeyIterator;
 use crate::metadata::PartitionMap;
 use crate::ops::{DataType, Datom, DatomOp, Entid, EntityRef, TxOp};
-use crate::schema::{Schema, Unique, ValueType};
+use crate::schema::{Schema, Unique, ValueType, DB_TX_ABORTED, DB_TX_COMMITTED};
+use crate::slate::DEFAULT_SCAN_OPTIONS;
+use crate::transaction::{TxBasis, TxKey};
 use crate::util::{concat_bytes, next_prefix};
 
 // ---------------------------------------------------------------------------
@@ -453,6 +457,87 @@ pub(crate) fn validate_allocated_entity_ids(
     }
 
     Ok(())
+}
+
+/// Look up a transaction's persisted outcome via the AVE index on `:db/txId`,
+/// reading `:db/txResult`/`:db/txError` from the tx entity.
+///
+/// Returns `None` if no tx entity exists: the tx is either not yet indexed or
+/// failed without persisting an outcome (technical abort, deserialize failure).
+pub(crate) async fn lookup_tx_completion<D>(
+    sdb: &D,
+    tx_key: TxKey,
+) -> Result<Option<TxCompletion>, Error>
+where
+    D: DbReadOps + Sync,
+{
+    // Keyed on tx_id only; the log assigns system_time together with tx_id.
+    let mut value_buf = Vec::new();
+    encode_datatype(&DataType::Long(tx_key.tx_id), &mut value_buf);
+    let ave_prefix = concat_bytes(&[
+        &[codec::AVE],
+        &encode_i64_bytes(crate::schema::DB_TX_ID),
+        &value_buf,
+    ]);
+    let mut iter = sdb
+        .scan_prefix_with_options(&ave_prefix, &DEFAULT_SCAN_OPTIONS)
+        .await?;
+    let mut tx_eid: Option<i64> = None;
+    while let Some(kv) = iter.next().await? {
+        let (_attribute, _value, entity, _tx_eid, op) = ave_key_to_parts(kv.key)?;
+        if op == codec::RETRACT {
+            continue;
+        }
+        match entity {
+            DataType::Long(eid) => {
+                tx_eid = Some(eid);
+                break;
+            }
+            other => bail!("Expected Long entity ID in AVE key, got {:?}", other),
+        }
+    }
+    let Some(tx_eid) = tx_eid else {
+        return Ok(None);
+    };
+
+    let mut entity_buf = Vec::new();
+    encode_datatype(&DataType::Long(tx_eid), &mut entity_buf);
+    let eav_prefix = concat_bytes(&[&[codec::EAV], &entity_buf]);
+    let mut iter = sdb
+        .scan_prefix_with_options(&eav_prefix, &DEFAULT_SCAN_OPTIONS)
+        .await?;
+    let mut tx_result: Option<i64> = None;
+    let mut tx_error: Option<String> = None;
+    while let Some(kv) = iter.next().await? {
+        let (_entity, attribute, value, _tx_eid, op) = eav_key_to_parts(kv.key)?;
+        if op == codec::RETRACT {
+            continue;
+        }
+        if attribute == crate::schema::DB_TX_RESULT {
+            if let DataType::Long(result) = value {
+                tx_result = Some(result);
+            }
+        }
+        if attribute == crate::schema::DB_TX_ERROR {
+            if let DataType::String(err) = value {
+                tx_error = Some(err);
+            }
+        }
+    }
+
+    let result = match tx_result {
+        Some(DB_TX_COMMITTED) => Ok(()),
+        Some(DB_TX_ABORTED) => Err(Arc::new(anyhow::anyhow!(
+            "{}",
+            tx_error.unwrap_or_else(|| format!("Transaction {} aborted", tx_key.tx_id))
+        ))),
+        Some(other) => bail!("Tx entity {tx_eid} has unknown :db/txResult {other}"),
+        None => bail!("Tx entity {tx_eid} missing :db/txResult"),
+    };
+    Ok(Some(TxCompletion {
+        basis: Some(TxBasis { tx_key, tx_eid }),
+        result,
+    }))
 }
 
 #[cfg(test)]
